requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-decouple
python-jose[cryptography]
bcrypt